LayoutTests/fast/multicol/balance-trailing-border.html

<!DOCTYPE html>
<script src="../../resources/js-test.js"></script>
<p>Both words below should be in the the right column.</p>
<div id="console"></div>
<div id="mc" style="-webkit-columns:2; columns:2; orphans:1; widows:1; -webkit-column-rule:1px solid; column-rule:1px solid; background:olive;">
    <div style="height:110px"></div>
    <div style="border:2px solid black; width:375px;">
        <div style="margin:0 10px; border:2px solid blue;">
            <span style="line-height:80px;">right<br>column</span>
        </div>
    </div>
</div>
<script>
    description("Tests that the bottom border at a column boundary is recorded properly for balancing");
    shouldBe("mc.offsetHeight", "168");
</script>
