third_party/blink/web_tests/fast/multicol/balance-trailing-border-expected.txt

Both words below should be in the the right column.

Tests that the bottom border at a column boundary is recorded properly for balancing

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS mc.offsetHeight is 168
PASS successfullyParsed is true

TEST COMPLETE

right
column
